bug fixes, filtres, new features

+++ handlers/personal_actions.py
@@ -34,9 +34,140 @@ async def cmd_ping_bot(message: types.Message):
 @dp.message_handler(IsOwner(), commands=['users'])
 @dp.async_task
 async def who_is_users(message: types.Message):
-    pre_msg = config['ALLOWED_IDS']
     await message.delete()
-    await message.answer(text=str(pre_msg)[1:-1])
+    pre_msg = config['ALLOWED_IDS']
+
+    if not(1 in pre_msg):
+        if pre_msg:
+            await message.answer(text=str(pre_msg)[1:-1])
+        else:
+            await message.answer(text='У вас нет пользователей.')
+    else:
+        await message.answer(text='У вас каждый - пользователь👏.')
+
+
+
+@dp.message_handler(IsOwner(), commands=['admins'])
+@dp.async_task
+async def who_is_admin(message: types.Message):
+    await message.delete()
+    pre_msg = config['OWNERS_IDS']
+
+    if not (1 in pre_msg):
+        if pre_msg:
+            await message.answer(text=str(pre_msg)[1:-1])
+        else:
+            await message.answer(text='У вас нет админов.')
+    else:
+        await message.answer(text='У вас каждый - админ😎👏.')
+
+
+
+@dp.message_handler(IsOwner(), commands=['addusers'])
+@dp.async_task
+async def add_users(message: types.Message):
+    try:
+        await message.delete()
+        text = (message.text[9:]).strip()
+        one = not (',' in text)
+        if text:
+            if one:
+                text = int(text.strip())
+                config['ALLOWED_IDS'].append(text)
+            else:
+                users = list(map(str.strip, text.split(',')))
+                new_users = [int(user) for user in users[::-1] if (user not in config['ALLOWED_IDS']) and user.isdigit()]
+                config['ALLOWED_IDS'].extend(new_users)
+
+            await confupdate()
+            await message.answer('Новые пользователи добавлены!')
+        else:
+            await message.answer('Для того, чтобы воспользоваться этой функцией, сделайте отступ и добавьте id пользователей через запятую.')
+    except:
+        await message.answer('Что-то не так... Проверьте что написали и повторите попытку.')
+
+
+
+@dp.message_handler(IsOwner(), commands=['addadmins'])
+@dp.async_task
+async def add_admins(message: types.Message):
+    try:
+        await message.delete()
+        text = (message.text[10:]).strip()
+        one = not (',' in text)
+        if text:
+            if one:
+                text = text.strip()
+                if text.isdigit():
+                    config['OWNERS_IDS'].append(int(text))
+                else:
+                    exit()
+            else:
+                admins = list(map(str.strip, text.split(',')))
+                new_users = [int(admin) for admin in admins[::-1] if (admin not in config['OWNERS_IDS']) and admin.isdigit()]
+                config['OWNERS_IDS'].extend(new_users)
+
+            await confupdate()
+            await message.answer('Новые админы добавлены!')
+        else:
+            await message.answer('Для того, чтобы воспользоваться этой функцией, сделайте отступ и добавьте id админов через запятую.')
+    except:
+        await message.answer('Что-то не так... Проверьте что написали и повторите попытку.')
+
+
+
+@dp.message_handler(IsOwner(), commands=['radmins'])
+@dp.async_task
+async def remome_admins(message: types.Message):
+    try:
+        await message.delete()
+        text = (message.text[8:]).strip()
+        one = not (',' in text)
+
+        if text:
+            if one:
+                if text.isdigit():
+                    config['OWNERS_IDS'].remove(int(text))
+                else:
+                    exit()
+            else:
+                [config['OWNERS_IDS'].remove(user) for user in[int(id.strip()) for id in text.split(',') if id.strip().isdigit()] if user in config['OWNERS_IDS']]
+
+
+            await confupdate()
+            await message.answer('Лишние id удалены!')
+        else:
+            await message.answer('Для того, чтобы воспользоваться этой функцией, сделайте отступ и добавьте id через запятую.')
+    except:
+        await message.answer('Что-то не так... Проверьте что написали и повторите попытку.')
+
+
+
+@dp.message_handler(IsOwner(), commands=['rusers'])
+@dp.async_task
+async def remome_users(message: types.Message):
+    try:
+        await message.delete()
+        text = (message.text[7:]).strip()
+        one = not (',' in text)
+
+        if text:
+            if one:
+                text = text.strip()
+                if text.isdigit():
+                    config['ALLOWED_IDS'].remove(int(text))
+                else:
+                    exit()
+            else:
+                [config['ALLOWED_IDS'].remove(user) for user in[int(id.strip()) for id in text.split(',') if id.strip().isdigit()] if user in config['ALLOWED_IDS']]
+
+
+            await confupdate()
+            await message.answer('Лишние id удалены!')
+        else:
+            await message.answer('Для того, чтобы воспользоваться этой функцией, сделайте отступ и добавьте id через запятую.')
+    except:
+        await message.answer('Что-то не так... Проверьте что написали и повторите попытку.')
 
 
 
@@ -59,13 +190,18 @@ async def update_bot(message: types.Message):
 @dp.async_task
 async def rollback_bot(message: types.Message):
     await message.delete()
-    m = await message.answer('Выполняю команду...')
-    run(['git', 'fetch'])
-    run(['git', 'reset', '--hard', 'origin/main'])
-    call(['git', 'reset', '--hard', 'HEAD~1'])
-    await m.edit_text('Бот вернулся к предыдущей версии, перезапуск...')
-    # перезапуск программы
-    await restart()
+    text = ((message.text)[8:]).strip()
+
+    if text:
+        await message.answer('Чтобы использовать эту команду, сделайте отступ от /gitback и укажите, на сколко версий назад вы хотите вернуться. Будьте осторожны! Откат на некоторые версии могут привести к неработоспособности бота.')
+    else:
+        m = await message.answer('Выполняю команду...')
+        run(['git', 'fetch'])
+        run(['git', 'reset', '--hard', 'origin/main'])
+        call(['git', 'reset', '--hard', f'HEAD~{text}'])
+        await m.edit_text('Бот вернулся к предыдущей версии, перезапуск...')
+        # перезапуск программы
+        await restart()
 
 
 
@@ -96,19 +232,53 @@ async def num_of_tokens(message: types.Message):
 
 
 
+@dp.message_handler(IsOwner(), commands=['printtokens'])
+@dp.async_task
+async def print_tokens(message: types.Message):
+    await message.delete()
+    await message.answer(text=str(api)[1:-1])
+
+
+
 @dp.message_handler(IsOwner(), commands=['addtokens'])
 @dp.async_task
 async def add_tokens(message: types.Message):
-    text = message.text[10:]
-    [config['OPENAI_TOKENS'].append(t) for t in [i.strip() for i in text.split(',')]]
-    await confupdate()
-    await message.answer('Новые токены добавлены!')
+    await message.delete()
+    text = (message.text[10:]).strip()
+    one = not (',' in text)
+    if text:
+        if one:
+            text = text.strip()
+            config['OPENAI_TOKENS'].append(text)
+        else:
+            tokens = list(map(str.strip, text.split(',')))
+            new_tokens = [token for token in tokens[::-1] if (token not in config['OPENAI_TOKENS']) and (token != '')]
+            config['OPENAI_TOKENS'].extend(new_tokens)
+
+        await confupdate()
+        await message.answer('Новые токены добавлены!')
+    else:
+        await message.answer('Для того, чтобы воспользоваться этой функцией, сделайте отступ и добавьте токены через запятую.')
 
 
 @dp.message_handler(IsOwner(), commands=['rtokens'])
 @dp.async_task
 async def remome_tokens(message: types.Message):
-    text = message.text[7:]
-    [config['OPENAI_TOKENS'].remove(t) for t in [i.strip() for i in text.split(',')]]
-    await confupdate()
-    await message.answer('Лишние токены удалены!')
+    await message.delete()
+    text = (message.text[8:]).strip()
+    one = not (',' in text)
+
+    if text:
+        if one:
+            text = text.strip()
+            config['OPENAI_TOKENS'].remove(text)
+        else:
+            for token in text.split(','):
+                token = token.strip()
+                if token in config['OPENAI_TOKENS'] and token != '':
+                    config['OPENAI_TOKENS'].remove(token)
+
+        await confupdate()
+        await message.answer('Лишние токены удалены!')
+    else:
+        await message.answer('Для того, чтобы воспользоваться этой функцией, сделайте отступ и добавьте токены через запятую.')
